ТЗ-8:
Создание абстрактного класса
